+++ lib/widgets/preview_description.dart
@@ -4,8 +4,9 @@ import 'package:simple_url_preview/widgets/empty_container.dart';
 class PreviewDescription extends StatelessWidget {
   final String _description;
   final Color _textColor;
+  final int _descriptionLines;
 
-  PreviewDescription(this._description, this._textColor);
+  PreviewDescription(this._description, this._textColor, this._descriptionLines);
 
   @override
   Widget build(BuildContext context) {
@@ -13,7 +14,7 @@ class PreviewDescription extends StatelessWidget {
       return Text(
         _description,
         textAlign: TextAlign.left,
-        maxLines: 3,
+        maxLines: _descriptionLines,
         style: TextStyle(
           fontSize: 14,
           color: _textColor,